rebase, pull file out, address comments
