+++ app/src/remote_server/mod.rs
@@ -11,6 +11,8 @@ pub use remote_server::*;
 #[cfg(not(target_family = "wasm"))]
 pub mod auth_context;
 #[cfg(not(target_family = "wasm"))]
+mod codebase_index_status;
+#[cfg(not(target_family = "wasm"))]
 pub mod diff_state_proto;
 #[cfg(not(target_family = "wasm"))]
 pub mod diff_state_tracker;
